newComment component and sendRef
